Move InfoDialog to container, include show logic
