+++ src/core/containers/InfoDialog/InfoDialog.scss
