+++ src/core/reducers/addons.js
@@ -26,6 +26,7 @@ export default function addon(state = initialState, action) {
       } else if (thisAddon.current_version && thisAddon.current_version.files.length > 0) {
         newState[key] = {
           ...thisAddon,
+          iconUrl: thisAddon.icon_url,
           installURL: thisAddon.current_version.files[0].url,
         };
       } else {
