+++ src/disco/containers/DiscoPane.js
@@ -9,8 +9,8 @@ import { camelCaseProps } from 'core/utils';
 import { loadEntities } from 'core/actions';
 import translate from 'core/i18n/translate';
 import tracking from 'core/tracking';
-import InfoDialog from 'core/components/InfoDialog';
 import { INSTALL_STATE } from 'core/constants';
+import InfoDialog from 'core/containers/InfoDialog';
 import { addChangeListeners } from 'core/addonManager';
 import {
   NAVIGATION_CATEGORY,
@@ -29,10 +29,8 @@ export class DiscoPaneBase extends React.Component {
     AddonComponent: PropTypes.func,
     handleGlobalEvent: PropTypes.func.isRequired,
     i18n: PropTypes.object.isRequired,
-    infoDialogData: PropTypes.object.isRequired,
     mozAddonManager: PropTypes.object,
     results: PropTypes.arrayOf(PropTypes.object).isRequired,
-    showInfoDialog: PropTypes.bool.isRequired,
     _addChangeListeners: PropTypes.func,
     _tracking: PropTypes.object,
   }
@@ -87,7 +85,7 @@ export class DiscoPaneBase extends React.Component {
   }
 
   render() {
-    const { results, i18n, AddonComponent, showInfoDialog, infoDialogData } = this.props;
+    const { results, i18n, AddonComponent } = this.props;
     const { showVideo } = this.state;
 
     return (
@@ -128,7 +126,7 @@ export class DiscoPaneBase extends React.Component {
             {i18n.gettext('See more add-ons!')}
           </a>
         </div>
-        {showInfoDialog === true ? <InfoDialog {...infoDialogData} /> : null}
+        <InfoDialog />
       </div>
     );
   }
@@ -154,8 +152,6 @@ export function loadDataIfNeeded({ store: { dispatch, getState } }) {
 export function mapStateToProps(state) {
   return {
     results: loadedAddons(state),
-    showInfoDialog: state.infoDialog.show,
-    infoDialogData: state.infoDialog.data,
   };
 }
